Update patch set 4
Patch Set 4:

The main reason for additional complexity of this new code is the fact that "peer chaincode package" uses the method of packaging the source to package the binary. So far, packaging was required to package only the source, and now we introduced the binary packagin instead. We should use specialized packaging for the binary, which could be actually quite simple, see generateDockerBuild() in dockerbuilder.go. Filename in cutil.WriteBytesToPackage() would be "chaincode" (as required by the golang docker).

Patch-set: 4
Label: Code-Review=0