[DEV-9546] Allow endorsers to create DB
Endorsers must be allowed to create a database
since, for private data, we can't rely on the committer
to create the DB since he may not have been part of the
transaction. The existing code handles race conditions
as a result of multiple peers attempting to create the
same DB, so we should be safe.

Change-Id: I0568e03b89a4785c2ddeb563723ac852dbbc68d4
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>