WIP
Reapplying [DEV-10853] "Committers using statecouchdb now store the CouchDB rev in"

Added statekeyindex.IndexReady interface that provides a channel through
which completion of the kvLedger's updates to the statekeyindex can be
notified to the statecouchdb batch committers.

Removed unnecessary check on couchdb pvt data store height for endorsers because
the endorser should always read from the database to get the height.

This reverts commit 1508312247de80a384cb61a27c30aca78c1c5285.

Change-Id: Icf9b963e2581ab78c53d045164e6217bb612311d
Signed-off-by: George Aristy <george.aristy@securekey.com>