[DEV-11326] Private data matching public hash not avail
A race condition exists where a new private data key has been
retrieved from the DB before the endorser had a chance to
commit the latest block. The private data is retrieved but
the committer hasn't yet committed the public hash of the key
and therefore the hash version is nil and the following error
is returned:

'Private data matching public hash version is not available.'

This patch fixes this problem by first retrieving the public
hash. If it doesn't exist then there's no need to load from
the private data store - simply return nil.

Change-Id: I91deaf6091e449d7f47902d751500e814c3742ec
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>