[DEV-11320] Fix for missing pvt data
Fix for a bug that occurs in an 'append-only' case.
Say that a collection is accessible only by org1
but the pvt data was added by org2 in an endorsement.
When org1 attempts to retrieve the data, it's not found.

This bug occurred because the code was checking the
list of endorsements to determine whether the pvt data
was missing. Since the endorsement happened only on
a peer that does not have access to the collection, the
list of endorsements was empty and the pvt data was not
considered to be missing.

This fix ensures that the pvt data is flagged as missing
even if the list of endorsements is empty.

Change-Id: Ia8293dbce28d198927cfa3f97ec05477d09e4cdc
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>