Fix deadlock when retrieving transient data
Intermittent deadlocks occur attempting to retrieve transient data. When a block is being committed, the committer attempts to acquire a write lock while the chaincode already has a read-lock. Then GetPrivateData attempts to acquire a read-lock to retrieve collection config data from the ledger. This commit adds an LSCC handler to the collection config retriever so that the collection config cache is populated directly from the data provided by the handler which eliminates the need to pull the config from the ledger.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>
Change-Id: I52539f2c39646c49f2ff06d988774e69e9b003ac