[BLOC-1836] Performance improvement for pulling private data
When private data is requested (pulled) from another peer, a response is
expected for every key that it requests or it times out after 5s. The
remote peer handles the request for each key by checking if it has the
data locally and if the requesting peer has access to the collection for
the key. If this criteria is not met then it does not produce a response
for that key and ends up replying with an empty array of private data.
This leaves the requesting peer waiting 5s before timing out.

This patch changes the behavior slightly in that the remote peer
responds to the request with an empty payload if it cannot fulfill the
request, so that the requesting peer knows immediately that the peer
does not have the data.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>
Change-Id: Ic7c822e3e598fc057fa8c043f0c308a18e5e8eed
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>